feat: adjusting barrels and file names
