+++ pages/_document.tsx
@@ -1,4 +1,4 @@
-import { ColorModeScript, theme } from '@chakra-ui/react'
+import { ColorModeScript, theme } from '@chakra-ui/react';
 import Document, {
 	Html,
 	Head,
@@ -6,14 +6,14 @@ import Document, {
 	NextScript,
 	DocumentContext,
 	DocumentInitialProps,
-} from 'next/document'
+} from 'next/document';
 
 class MyDocument extends Document {
 	static async getInitialProps(
 		ctx: DocumentContext
 	): Promise<DocumentInitialProps> {
-		const initialProps = await Document.getInitialProps(ctx)
-		return { ...initialProps }
+		const initialProps = await Document.getInitialProps(ctx);
+		return { ...initialProps };
 	}
 
 	render() {
@@ -33,8 +33,8 @@ class MyDocument extends Document {
 					<NextScript />
 				</body>
 			</Html>
-		)
+		);
 	}
 }
 
-export default MyDocument
+export default MyDocument;
