+++ pages/api/hello.ts
@@ -1,13 +1,13 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
-import type { NextApiRequest, NextApiResponse } from 'next'
+import type { NextApiRequest, NextApiResponse } from 'next';
 
 type Data = {
-	name: string
-}
+	name: string;
+};
 
 export default function handler(
 	req: NextApiRequest,
 	res: NextApiResponse<Data>
 ) {
-	res.status(200).json({ name: 'John Doe' })
+	res.status(200).json({ name: 'John Doe' });
 }
