+++ prettier.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-	semi: false,
+	semi: true,
 	singleQuote: true,
 	arrowParens: 'avoid',
-};
+}
