+++ styles/index.ts
@@ -1,1 +1,1 @@
-export * from './theme'
+export * from './theme';
