+++ styles/theme.ts
@@ -1,9 +1,9 @@
-import { extendTheme, ThemeConfig } from '@chakra-ui/react'
+import { extendTheme, ThemeConfig } from '@chakra-ui/react';
 
 const config: ThemeConfig = {
 	initialColorMode: 'dark',
 	useSystemColorMode: false,
-}
+};
 
 export const theme = extendTheme({
 	config,
@@ -21,4 +21,4 @@ export const theme = extendTheme({
 		body: 'Roboto',
 		mono: 'Roboto',
 	},
-})
+});
